Moved some documents tests to correct place
